Change payout reward type from String to Long

+++ altchain-sdk/src/main/kotlin/org/veriblock/sdk/alt/model/SecurityInheritingTransaction.kt
@@ -7,6 +7,6 @@ data class SecurityInheritingTransaction(
 )
 
 data class SecurityInheritingTransactionVout(
-    val value: Double,
+    val value: Long,
     val addressHex: String
 )
